requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
bcrypt==4.1.1
PyJWT==2.12.1
requests==2.31.0
validators==0.22.0
tldextract==5.3.1
scikit-learn==1.3.1
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2
python-multipart==0.0.6
slowapi==0.1.8
